# fix: Memoize SidebarMatch to avoid rerendering as results arrive