--- a/src/ts/components/Markdown.tsx
+++ b/src/ts/components/Markdown.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { getHtmlFromMarkdown } from "../utils/dom";
+
+const Markdown = ({ markdown }: { markdown: string }) => (
+  <div
+    dangerouslySetInnerHTML={{
+      __html: getHtmlFromMarkdown(markdown)
+    }}
+  />
+);
+
+export default Markdown;