## What does this change?

Memoizes the `SidebarMatch` component, which is continually rerendered as our state changes.

This results in a significant improvement in our JS calc times. Timings taken when rechecking an article of 4,540 (to better isolate unneeded rerenders), averaged over 5 tries:

| Before | After |
|-|-|
| 6252ms | 4234ms |

## How to test

This is a no-op for features, but the impact is noticeable when using CPU throttling in e.g. Chrome.